net: rbridge: insert VLAN in the inner header on ingress
Handle VLAN as payload when encapsulating native frames: upon ingress,
the native frame's VLAN is to be set in the inner header. References
to this VLAN must also be dropped from the sk_buff, otherwise it would
be restored in the outer header upon xmit.

(this patch was in fact already present in the VNT which is not merged
at the moment, but we forgot to import it in the TRILL part)

[slightly edit comment]
Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>
Signed-off-by: William Dauchy <william@gandi.net>
Signed-off-by: Ahmed Amamou <ahmed@gandi.net>