net: rbridge: fix local host frame forwarding
when trill is enabled we used to consider as local packet all packet with
destination mac address that is local this makes packet destined to
FE:FF:FF:FF:FF:FF considered as host packets.
such packets should not be forwarded by switches to nodes.
this also will cause nodes with 2 interfaces attached to the bridge to accept
packet with a destination equal to any of these two interfaces address.
reporting to bridge code correct behavious should be considering only packets
a destination equal to bridge mac address as host packets

Signed-off-by: Ahmed Amamou <ahmed@gandi.net>